% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likBivNorm_function.R
\name{bothUncen.N}
\alias{bothUncen.N}
\title{Log Likelihood Function Both Uncensored Normal}
\usage{
bothUncen.N(thetaI, cenData)
}
\arguments{
\item{thetaI}{vector of parameters to optimize untransformed such that c( xmu, ymu, xsig, ysig, r).}

\item{cenData}{matrix of data in which first two columns hold data values for x and y and the second two
columns hold flags for censoring such that 1 implies censored and 0 implies not censored.}
}
\value{
Vector whose sum would be log likelihood function if completely uncensored.
}
\description{
This is the log likelihood function when both x and y are uncensored for bivariate, normally distributed data.
}

