% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likBiv.T_functions.R
\name{bothUncen.T}
\alias{bothUncen.T}
\title{Log Likelihood Function Both Uncensored T}
\usage{
bothUncen.T(thetaIS, cenData, df)
}
\arguments{
\item{thetaIS}{vector of parameters to optimize, as scale parameters and untransformed.}

\item{cenData}{matrix of data in which first two columns hold data values for x and y and the second two
columns hold flags for censoring such that 1 implies censored and 0 implies not censored.}

\item{df}{Degrees of Freedom must be greater than 3 and an integer.}
}
\value{
Vector whose sum would be log likelihood function if completely uncensored.
}
\description{
This is the log likelihood function when both x and y are uncensored for bivariate T distributed data.
}

