% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScaleMat_function.R
\name{buildScaleMat}
\alias{buildScaleMat}
\title{Build Scale Matrix}
\usage{
buildScaleMat(xsd, ysd, r, df)
}
\arguments{
\item{xsd}{A positive number representing the standard deviation of x.}

\item{ysd}{A positive number representing the standard deviation of x.}

\item{r}{A number between -1 and 1 representing the correlation of x and y.}

\item{df}{Integer greater than 3, representing degrees of freedom with df=Inf implying normal.}
}
\value{
Returns a 2x2 matrix oftened used to give scaling parameters to rmt.
}
\description{
This function builds a 2x2 scale matrix given xsd, ysd, corr, and df.
}
\examples{
buildScaleMat( 1,1,0,Inf) #normal 
buildScaleMat( 1,1,0,4) #t with df =4

}

