% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_functions.R
\name{censorData}
\alias{censorData}
\title{Left Censor Data}
\usage{
censorData(uncenData, cenLevelVec)
}
\arguments{
\item{uncenData}{Data matrix of any size.}

\item{cenLevelVec}{vector of length equal to rows of data, indicating censoring level requested in decimals.}
}
\value{
Matrix such that the first half of the columns are the censored data and the 
        second haf are the isCensored flags with 1 indicating censoring, where censored data has a value
        equal to the limit of detection.
}
\description{
This function will take a data set and artificially left censor it to a request level of censoring.
}
\details{
The function sets the limit of detection at the value which the requested amount of data is below.
         All values below this limit of detection are replaced with the limit of detection.
}
\examples{
censorData(uncenData = myData, cenLevelVec =c(.2,.2))

}

