% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censor_function.R
\name{censorDifferently}
\alias{censorDifferently}
\title{Alter Censoring Method}
\usage{
censorDifferently(cenData, replacement = LOD/sqrt(2))
}
\arguments{
\item{cenData}{Data matrix such that the first half of the columns are the censored data and the 
second haf are the isCensored flags with 1 indicating censoring, where censored data has a value
equal to the limit of detection.}

\item{replacement}{The value request to fill censored values in the censored data matrix. 
By Defualt equals LOD divided by sqrt2 as this is a common method.}
}
\value{
Matrix such that the first half of the columns are the censored data and the 
        second haf are the isCensored flags with 1 indicating censoring, where censored data has a value
        equal to replace parameter.
}
\description{
This function will take already censored data and replace LOD with LOD/sqrt2 or another value
}
\details{
To change the censored values to LOD/sqrt2 the data passed to the function must be censored
         such that censored values = LOD. If LOD is not availible, replacement should only equal numerical values.
}
\examples{

}

