% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_functions.R
\name{genData.N}
\alias{genData.N}
\title{Generate Bivariate Norm Data}
\usage{
genData.N(n, xmu, ymu, xsd, ysd, r)
}
\arguments{
\item{n}{number of pairs of data to generate, number of rows in data set, or sample size}

\item{xmu}{x variable's mean}

\item{ymu}{y variable's mean}

\item{xsd}{x variable's standard deviation}

\item{ysd}{y varaible's standard deviation}

\item{r}{correlation between the x and y variable}
}
\value{
data frame of uncensored data with n rows and 2 columns
}
\description{
This function will generate an uncensored dataset distributed with bivaraite normal
}
\examples{

}

