% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData_function.R
\name{genData}
\alias{genData}
\title{Generate Bivariate Data}
\usage{
genData(n, locVec, scaleMat, df)
}
\arguments{
\item{n}{An integer, representing the number of pairs of data to generate, number of rows in data set, or sample size.}

\item{locVec}{A vector of length 2, representing the location parameter (equal to the mean vector when df>1).}

\item{scaleMat}{A symmetric positive-definite 2x2 matrix representing the scale matrix of the distribution, such that S*df/(df-2) is the variance-covariance matrix when df>2}

\item{df}{Integer greater than 3, representing degrees of freedom with df=Inf implying normal.}
}
\value{
Returns a data set matrix with nrow = n and ncol= 2.
}
\description{
This function will generate an uncensored dataset distributed bivaraite normal or t.
}
\details{
It is suggested to use buildScaleMat to generate the scaleMat parameter.
}
\examples{
xmu = 0
ymu = 0
xsd = 1
ysd = 1
r = 0
df = Inf #normal
scaleMat <- buildScaleMat( xsd, ysd, r, df)
genData(10, c(xmu, ymu), scaleMat, df)

}

