% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likBivNorm_function.R
\name{likBivNorm}
\alias{likBivNorm}
\title{Likelihood Function for Normal Left Censored Data}
\usage{
likBivNorm(thetaT, cenData)
}
\arguments{
\item{thetaT}{transformed vector of parameters to optimize such that c(xmu, ymu, log(xsig), log(ysig), tanh(r)).}

\item{cenData}{matrix of data in which first two columns hold data values for x and y and the second two
columns hold flags for censoring such that 1 implies censored and 0 implies not censored.}
}
\value{
calculated maximum likelihood function for optimazation using optim.
}
\description{
This function is the maximum likelihood function of bivariate, normally distributed, left censored data.
}
\examples{

}

