% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likBivT_Function.R
\name{likBivT}
\alias{likBivT}
\title{Likelihood Function for T Left Censored Data}
\usage{
likBivT(thetaT, cenData, df)
}
\arguments{
\item{thetaT}{transformed vector of parameters to optimize such that c(xmu, ymu, log(xsig), log(ysig), tanh(r)).}

\item{cenData}{matrix of data in which first two columns hold data values for x and y and the second two
columns hold flags for censoring such that 1 implies censored and 0 implies not censored.}

\item{df}{Degrees of Freedom must be greater than 3 and an integer.}
}
\value{
calculated maximum likelihood function for optimazation using optim.
}
\description{
This function is the maximum likelihood function of bivariate T distributed, left censored data.
}

