% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimStartGuess_function.R
\name{optimStartGuess}
\alias{optimStartGuess}
\title{Optim Starting Guess}
\usage{
optimStartGuess(cenData)
}
\arguments{
\item{cenData}{matrix of data in which first two columns hold data values for x and y and the second two
columns hold flags for censoring such that 1 implies censored and 0 implies not censored.}

\item{thetaT}{transformed vector of parameters to optimize such that c(xmu, ymu, log(xsig), log(ysig), tanh(r)).}
}
\value{
guesstheta vector of length 5, guess for each xmu, ymu, xsd, ysd, rho
}
\description{
This function picks the reasonable starting guess for optim, which is the mean, sd, and corr of the data ignoring censoring.
}

