% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_functions.R
\name{transformTheta}
\alias{transformTheta}
\title{Transform Theta}
\usage{
transformTheta(thetaI)
}
\arguments{
\item{thetaI}{vector of length 5, such that c(mux, muy, sigx, sigy, r). I stands for inverted/orignal.}
}
\value{
thetaT vector of length 5, such that c( mux, muy, log(sigx), log(sigy), atanh(r))
}
\description{
This function transforms theta to impose proper/logical bounds for std dev and rho
}

