% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_functions.R
\name{transformThetaInv}
\alias{transformThetaInv}
\title{Transform Theta Inverse}
\usage{
transformThetaInv(thetaT)
}
\arguments{
\item{thetaT}{vector of length 5, such that c( mux, muy, log(sigx), log(sigy), atanh(r))}
}
\value{
thetaI vector of length 5, such that c(mux, muy, sigx, sigy, r). I stands for inverted/orignal.
}
\description{
This function undoes the transformation to theta from the transfromTheta function.
}

