% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likBivNorm_function.R
\name{xOnlyCen.N}
\alias{xOnlyCen.N}
\title{Log Likelihood Function x Only Censored Normal}
\usage{
xOnlyCen.N(thetaI, cenData)
}
\arguments{
\item{thetaI}{vector of parameters to optimize untransformed such that c( xmu, ymu, xsig, ysig, r).}

\item{cenData}{matrix of data in which first two columns hold data values for x and y and the second two
columns hold flags for censoring such that 1 implies censored and 0 implies not censored.}
}
\value{
Vector whose to be summed for this portion of the log liklihood function.
}
\description{
This is the portion of the log likelihood function when the x variable is cesnored and
   the y varaible isn't censored for bivariate, normally distributed data.
}
\examples{

}

