% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likBivT_Function.R
\name{yCenOnly.T}
\alias{yCenOnly.T}
\title{Log Likelihood Function y Only Censored T}
\usage{
yCenOnly.T(thetaIS, cenData, df)
}
\arguments{
\item{thetaIS}{vector of parameters to optimize, as scale parameters and untransformed.}

\item{cenData}{matrix of data in which first two columns hold data values for x and y and the second two
columns hold flags for censoring such that 1 implies censored and 0 implies not censored.}

\item{df}{Degrees of Freedom must be greater than 3 and an integer.}
}
\value{
Vector whose to be summed for this portion of the log liklihood function.
}
\description{
This is the portion of the log likelihood function when the y variable is cesnored and
   the x varaible isn't censored for bivariate t distributed data.
}

