% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScaleMat_function.R
\name{buildScaleMat}
\alias{buildScaleMat}
\title{Build Scale Matrix}
\usage{
buildScaleMat(xsd, ysd, r, df)
}
\arguments{
\item{xsd}{standard deviation of x.}

\item{ysd}{standard deviation of y.}

\item{r}{correlation of x and y.}

\item{df}{Degrees of Freedom must be greater than 3 and an integer or Inf for normal.}
}
\value{
2x2 matrix
}
\description{
This function builds a 2x2 scale matrix given xsd, ysd, corr, and df.
}
\examples{
buildScaleMat( 1,1,0,Inf)

}

