% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultGuess_function.R
\name{defaultGuess}
\alias{defaultGuess}
\title{Default Guess}
\usage{
defaultGuess(cenData)
}
\arguments{
\item{cenData}{matrix of data in which first two columns hold data values for x and y and the second two
columns hold flags for censoring such that 1 implies censored and 0 implies not censored.}
}
\value{
guesstheta vector of length 5, guess for each xmu, ymu, xsd, ysd, rho
}
\description{
This function picksa reasonable starting guess for optim, which is the mean, sd, and corr of the data ignoring censoring.
}

