% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData_function.R
\name{genData}
\alias{genData}
\title{Generate Bivariate Data}
\usage{
genData(n, locVec, scaleMat, df)
}
\arguments{
\item{n}{number of pairs of data to generate, number of rows in data set, or sample size.}

\item{locVec}{a vector of length d, representing the location parameter (equal to the mean vector when df>1).}

\item{scaleMat}{a symmetric positive-definite matrix representing the scale matrix of the distribution, such that S*df/(df-2) is the variance-covariance matrix when df>2}

\item{df}{Degrees of Freedom must be greater than 3 and an integer or Inf implying normal.}
}
\value{
data set matrix with nrow = n and ncol= d.
}
\description{
This function will generate an uncensored dataset distributed with bivaraite normal or t.
}
\examples{
genData(10, c(0,0), buildScaleMat(1,1,0,Inf), Inf)
}

